+++ app/components/layout/admin_header_component.html.erb
@@ -34,13 +34,13 @@
         <% end %>
       </div>
 
-      <% if show_admin_menu?(current_user) || namespace != "management" %>
+      <% if show_admin_menu?(user) || namespace != "management" %>
         <div id="responsive_menu">
           <div class="top-bar-right">
             <ul class="menu" data-responsive-menu="medium-dropdown">
-              <%= render "shared/admin_login_items", current_user: current_user %>
-              <%= render "layouts/notification_item", current_user: current_user %>
-              <%= render "devise/menu/login_items", current_user: current_user %>
+              <%= render "shared/admin_login_items", current_user: user %>
+              <%= render "layouts/notification_item", current_user: user %>
+              <%= render "devise/menu/login_items", current_user: user %>
             </ul>
           </div>
         </div>
