+++ app/helpers/admin_helper.rb
@@ -3,16 +3,6 @@ module AdminHelper
     "/#{namespace}"
   end
 
-  def namespaced_header_title
-    if namespace == "moderation/budgets"
-      t("moderation.header.title")
-    elsif namespace == "management"
-      t("management.dashboard.index.title")
-    else
-      t("#{namespace}.header.title")
-    end
-  end
-
   def official_level_options
     options = [["", 0]]
     (1..5).each do |i|
